Rename settings module to config
to limit confusion with the settings object included in settings module

# Conflicts:
#	app/endpoints/auth.py
#	tests/commons.py
