+++ app/endpoints/users.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Body, Depends, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core import security
-from app.core.settings import settings
+from app.core.config import settings
 from app.cruds import cruds_groups, cruds_users
 from app.dependencies import get_current_user, get_db
 from app.models import models_core
